+++ conversation/go/sdk/dapr.yaml
@@ -2,6 +2,6 @@ version: 1
 apps:
   - appDirPath: ./conversation/
     appID: conversation
-    appPort: 6300
-    daprHTTPPort: 6380
+    appPort: 3500
+    daprHTTPPort: 3501
     command: ["go", "run", "."]
